feat(Decks): add adaptative methods for getters if user is authenticated

+++ backend/app/Http/Controllers/DeckController.php
@@ -2,37 +2,66 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\DeckVisibility;
 use App\Http\Controllers\FlashcardController;
 use App\Http\Requests\StoreDeckRequest;
 use App\Http\Requests\UpdateDeckRequest;
 use App\Http\Resources\DeckCollection;
 use App\Http\Resources\DeckResource;
 use App\Models\Deck;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DeckController
 {
     /**
-     * Display a listing of the resource.
+     * Display a listing of the resource by page.
      */
-    public function getDecksByUser()
+    public function getDecksByPage(Request $request)
     {
         try {
-            // return new DeckCollection(Deck::where("user_id", auth()->user()->id)->get());
-            return response()->json(new DeckCollection(Deck::with("flashcards")->get()), 200);
+            $numberPerPage = 10;
+            $myDecks = $request->has("myDecks");
+            $decks = Deck::with("flashcards");
+
+            if ($myDecks) {
+                $user = Auth::guard('sanctum')->user();
+                if (!$user) {
+                    return response()->json(["message" => "Unauthorized"], 401);
+                }
+
+                $decks = $decks->where("user_id", $user->id);
+            } else {
+                $decks = $decks->where("visibility", DeckVisibility::PUBLIC ->value);
+            }
+
+            return response()->json(new DeckCollection($decks->paginate($numberPerPage)), 200);
         } catch (\Exception $e) {
             return response()->json(["error" => $e->getMessage()], 400);
         }
     }
 
     /**
-     * Display a listing of the resource.
+     * Display one item of the resource.
      */
-    public function getDeckById(int $id)
+    public function getDeckById(int $id, Request $request)
     {
         try {
             $deck = Deck::with("flashcards")->find($id);
             if (!$deck) {
                 return response()->json(["message" => "Deck not found"], 404);
+            }
+
+            $user = Auth::guard('sanctum')->user();
+
+            if ($deck->visibility == "Private") {
+                if (!$user) {
+                    return response()->json(["message" => $user->id], 401);
+                }
+
+                if ($user->id != $deck->user_id) {
+                    return response()->json(["message" => "Forbidden"], 403);
+                }
             }
 
             return response()->json(new DeckResource($deck), 200);
@@ -47,10 +76,13 @@ class DeckController
     public function createDeck(StoreDeckRequest $request, FlashcardController $flashcardController)
     {
         try {
+            $user = $request->user();
+
             $deck = Deck::create([
                 "name" => $request->name,
-                "visibility" => $request->visibility ? $request->visibility : "Public",
+                "visibility" => $request->visibility ? $request->visibility : DeckVisibility::PUBLIC ->value,
                 "likes" => 0,
+                "user_id" => $user->id,
             ]);
 
             foreach ($request->flashcards as $flashcard) {
@@ -72,6 +104,11 @@ class DeckController
             $deck = Deck::find($id);
             if (!$deck) {
                 return response()->json(["message" => "Deck not found"], 404);
+            }
+
+            $user = $request->user();
+            if ($deck->user_id != $user->id) {
+                return response()->json(["message" => "Forbidden"], 403);
             }
 
             $deck->update([
@@ -97,12 +134,17 @@ class DeckController
     /**
      * Remove the specified resource from storage.
      */
-    public function deleteDeckById(int $id)
+    public function deleteDeckById(int $id, Request $request)
     {
         try {
             $deck = Deck::find($id);
             if (!$deck) {
                 return response()->json(["message" => "Deck not found"], 404);
+            }
+
+            $user = $request->user();
+            if ($deck->user_id != $user->id) {
+                return response()->json(["message" => "Forbidden"], 403);
             }
 
             $deck->delete();
